refactor(1682 item2): promote _branch_tree → public branch_tree module
#2 PR2 follow-up (#1682): chainlit/rewind_actions.py top-level-imports across the
package boundary from the PRIVATE-named `reyn.tui.widgets._branch_tree` module —
but that module is genuinely a PUBLIC helper (public __all__: build_branch_tree_rows
/ ROW_* / selectable_rows; documented "one tree-grouping source across TUI + web,
lead-approved cross-surface reuse"). The leading `_` was a misnomer.

Fix = drop the `_` (lead's primary framing): git-mv `_branch_tree.py` →
`branch_tree.py` + repoint the ~10 importers (tui app/rewind_menu/conversation
relative + chainlit cross-surface + ~6 tests) `widgets._branch_tree` →
`widgets.branch_tree`. Eliminates the cross-package private-symbol import at the
source (vs a re-export shim — the surface is small, so a direct rename is cleaner,
consistent with the item-1 shim-vs-rewrite reasoning).

Verified: branch_tree (public) + all importers + chainlit import OK; 24 branch-tree
tests pass; ruff src tests clean.

Refs #1682 (#2 PR2 follow-up).
