+++ tests/cli/test_rewind_menu_tree_2b.py
@@ -21,7 +21,7 @@ if str(_SRC) not in sys.path:
 
 from reyn.tui.app import ReynTUIApp
 from reyn.tui.widgets import ConversationView
-from reyn.tui.widgets._branch_tree import build_branch_tree_rows
+from reyn.tui.widgets.branch_tree import build_branch_tree_rows
 from reyn.tui.widgets.rewind_menu import _MAX_VISIBLE, RewindMenuWidget
