+++ tests/test_anchor_text_1547.py
@@ -17,7 +17,7 @@ from reyn.events.agent_snapshot import AgentSnapshot
 from reyn.events.anchor_store import AnchorStore, truncate_anchor
 from reyn.events.snapshot_generations import SnapshotGenerationStore
 from reyn.events.state_log import StateLog
-from reyn.tui.widgets._branch_tree import build_branch_tree_rows
+from reyn.tui.widgets.branch_tree import build_branch_tree_rows
 from reyn.tui.widgets.rewind_menu import RewindMenuWidget
 
 # ── truncate_anchor ──────────────────────────────────────────────────────────
