+++ tests/test_branch_tree_2b.py
@@ -10,7 +10,7 @@ No mocks, no app — pure functions over synthetic contract-shape data.
 """
 from __future__ import annotations
 
-from reyn.tui.widgets._branch_tree import (
+from reyn.tui.widgets.branch_tree import (
     ROW_CHECKPOINT,
     ROW_HEADER,
     build_branch_tree_rows,
